fix(security): prevent secret leakage, harden supply chain and headers
Close a high-severity leak where the TMDB API key and Slack webhook URL
could reach unauthenticated /sync and /notifications/test response bodies
(and logs) via reqwest error strings that embed the request URL. Strip the
URL centrally in From<reqwest::Error> via without_url(), and add
AppError::client_message() so internal error variants never serialize into
a response body. (#70)

Supply chain: pin every GitHub Action to a commit SHA and add
helpers:pinGitHubActionDigests (#71); build the frontend image from the
audited lockfile on node:24-slim with `npm ci --ignore-scripts` and add
`cargo build --locked` (#72); scope ci.yml permissions, pin the Socket CLI,
narrow `secrets: inherit`, and give Renovate digest auto-merge a soak (#73).

Hardening: add CSP + X-Content-Type-Options/X-Frame-Options/Referrer-Policy
response headers (#74); reject NOTIFICATION_CHECK_INTERVAL_MINUTES=0 instead
of panicking, widen notification season/episode to i64, cap TMDB response
size, and quote/encode TMDB-derived URLs in the frontend (#75).

+++ backend/src/logic/notifications.rs
@@ -74,8 +74,8 @@ pub async fn check_and_notify(
 fn build_event(ep: &NotifiableEpisode) -> NotificationEvent {
     NotificationEvent::EpisodeAiringToday {
         show_name: ep.show_name.clone(),
-        season: ep.season_number as u16,
-        episode: ep.episode_number as u16,
+        season: ep.season_number,
+        episode: ep.episode_number,
         title: ep.episode_name.clone().unwrap_or_default(),
         where_to_watch: ep.watch_providers.clone(),
     }
